refactor(merge): merged in fifo queue update

+++ Sources/TUSKit/Files.swift
@@ -89,9 +89,7 @@ final class Files {
             }
         }
         
-        try makeDirectoryIfNeeded(nil)
-        
-        try self.convertOldCacheToMMKV()
+       try self.convertOldCacheToMMKV()
     }
     
     static private var documentsDirectory: URL {
@@ -101,39 +99,45 @@ final class Files {
     /// Moves any files from TUS/background to TUS and inserts metadata into MMKV
     private func convertOldCacheToMMKV() throws {
         try fileQueue.sync {
-            let uuidDirs = try self.contentsOfDirectory(directory: type(of: self).documentsDirectory.appendingPathComponent("TUS/background"))
             
-            let decoder = PropertyListDecoder()
-                        
-            try uuidDirs.compactMap { uuidDir in
-                let uuidDirContents = try contentsOfDirectory(directory: uuidDir)
-                let metaDataUrls = uuidDirContents.filter{ $0.pathExtension == "plist" }
-                if(metaDataUrls.isEmpty) {
-                    try FileManager.default.removeItem(at: uuidDir)
-                } else {
-                    let metaDataUrl = metaDataUrls[0]
-                    if let data = try? Data(contentsOf: metaDataUrl) {
-                        let metaData = try? decoder.decode(UploadMetadata.self, from: data)
-                        metaData?.fileDir = metaDataUrl.deletingLastPathComponent()
-                        
-                        guard let metaData = metaData else {
-                            return
-                        }
-                        
-                        // Insert into MMKV
-                        try self.encodeAndStore(metaData: metaData)
-                        
-                        // Move file to TUS/ directory
-                        try self.copyAndChunk(from: metaDataUrl, id: metaData.id, chunkSize: -1)
-                        
-                        // Add upload manifest to queue
-                        guard let uploadManifestId = metaData.context?[UPLOAD_MANIFEST_METADATA_KEY] else {
-                            return
-                        }
-                        self.addFileToUploadManifest(uploadManifestId, uuid: metaData.id)
-                        
-                        // Delete directory in TUS/background
+            let doesExist = FileManager.default.fileExists(atPath: type(of: self).documentsDirectory.appendingPathComponent("TUS/background").path, isDirectory: nil)
+            
+            if doesExist {
+                let uuidDirs = try self.contentsOfDirectory(directory: type(of: self).documentsDirectory.appendingPathComponent("TUS/background"))
+                
+                let decoder = PropertyListDecoder()
+                
+                try uuidDirs.compactMap { uuidDir in
+                    let uuidDirContents = try contentsOfDirectory(directory: uuidDir)
+                    let metaDataUrls = uuidDirContents.filter{ $0.pathExtension == "plist" }
+                    if(metaDataUrls.isEmpty) {
                         try FileManager.default.removeItem(at: uuidDir)
+                    } else {
+                        let metaDataUrl = metaDataUrls[0]
+                        print("URL: \(metaDataUrl)")
+                        if let data = try? Data(contentsOf: metaDataUrl) {
+                            let metaData = try? decoder.decode(UploadMetadata.self, from: data)
+                            metaData?.fileDir = metaDataUrl.deletingLastPathComponent()
+                            
+                            guard let metaData = metaData else {
+                                return
+                            }
+                            
+                            // Insert into MMKV
+                            try self.encodeAndStore(metaData: metaData)
+                            
+                            // Move file to TUS/ directory
+                            try self.copyAndChunk(from: metaDataUrl, id: metaData.id, chunkSize: -1)
+                            
+                            // Add upload manifest to queue
+                            guard let uploadManifestId = metaData.context?[UPLOAD_MANIFEST_METADATA_KEY] else {
+                                return
+                            }
+                            self.addFileToUploadManifest(uploadManifestId, uuid: metaData.id)
+                            
+                            // Delete directory in TUS/background
+                            try FileManager.default.removeItem(at: uuidDir)
+                        }
                     }
                 }
             }
